// back-end/src/users/dto/user.dto.interface.ts
export interface IUserDto {
    email?: string;
    name: string;
    firstname?: string;
    birthday?: string; // Use string for birthday
    tel_num?: string;
    tel_medic?: string;
    tel_emergency?: string;
    weight?: string; // Use string for weight
    license?: string;
    date_subscribe?: string; // Use string for date_subscribe
    date_payment?: string;
    date_end_pay?: string;
    avatar?: string;
}